move client to the lib folder
