--- src/clients/index.ts
@@ -1,3 +1,0 @@
-export { APIClient } from './APIClient';
-export { Client, ClientFactory } from './ClientFactory';
-export { ClientInterface } from './ClientInterface';
