--- src/clients/local/localSampleContext.ts
@@ -1,11 +1,0 @@
-export default {
-  apis: [
-    {
-      name: 'todos',
-      url: 'https://jsonplaceholder.cypress.io/todos',
-    },
-  ],
-  models: {
-    primes: [1, 3, 5, 7, 11, 13],
-  },
-};
